Ensure that we can accept multiple options
Pattern matching in Elixir functions does not allow you to match on part of a
keyword list, as it does with maps. Previously, the function looked for
`[nested: true]` which would not match `[nested: true, input_opts: [class:
"button"]]`. This removes the use of pattern matching and uses a conditional to
avoid that problem.